+++ tests/test_processing.py
@@ -307,12 +307,20 @@ class TestNormalisation(unittest.TestCase):
     def test_is_undoable(self):
         self.assertTrue(self.processing.undoable)
 
+    def test_with_unknown_kind_raises(self):
+        self.processing.parameters["kind"] = 'foo'
+        with self.assertRaisesRegex(ValueError, 'not recognised'):
+            self.dataset.process(self.processing)
+
     def test_normalise_to_maximum(self):
         self.processing.parameters["kind"] = 'maximum'
         self.dataset.process(self.processing)
         self.assertEqual(1, self.dataset.data.data.max())
 
     def test_normalise_to_minimum(self):
+        # Asymmetric data with maximum < minimum
+        self.dataset.data.data = \
+            np.sin(np.linspace(0.6*np.pi, 2.4*np.pi, num=500))*2
         self.processing.parameters["kind"] = 'minimum'
         self.dataset.process(self.processing)
         self.assertAlmostEqual(-1, self.dataset.data.data.min(), 4)